requirements.txt
requests==2.32.3
pandas==2.2.3
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytest==8.3.3
matplotlib==3.10.1
